apps/testAPI/requirements_test.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-decouple==3.4
requests==2.25.1
urllib3==1.26.4
Werkzeug==2.0.1
